Move filter_connectivity_event (p2p)
Move filter_connectivity_event to testing_utils module.
