+++ p2p/src/peer_manager/helpers.rs
@@ -60,27 +60,3 @@ where
 
     (address, peer_info1, peer_info2)
 }
-
-/// Return first event that is accepted by predicate.
-///
-/// Used to skip events that are not of interest or that are different between backends
-/// (for example ConnectivityEvent::Discovered).
-/// Can be used in tests only, will panic in case of errors.
-pub async fn filter_connectivity_event<T, F>(
-    conn: &mut T::ConnectivityHandle,
-    predicate: F,
-) -> crate::Result<ConnectivityEvent<T>>
-where
-    T: NetworkingService,
-    T::ConnectivityHandle: ConnectivityService<T>,
-    F: Fn(&crate::Result<ConnectivityEvent<T>>) -> bool,
-{
-    loop {
-        let result = timeout(Duration::from_secs(10), conn.poll_next())
-            .await
-            .expect("unexpected timeout receiving connectivity event");
-        if predicate(&result) {
-            return result;
-        }
-    }
-}
